More Changes
The project can build now.  Still needs more work.
