+++ exi2xml/CMakeLists.txt
@@ -1,43 +1,36 @@
-cmake_minimum_required (VERSION 3.8)
-project(exi2xml)
+cmake_minimum_required (VERSION 3.21)
 
-# specify the C++ standard
-set(CMAKE_CXX_STANDARD 14)
-set(CMAKE_CXX_STANDARD_REQUIRED True)
+project(exi2xml
+	VERSION 0.2.0
+	DESCRIPTION "EXI to XML writer library"
+	LANGUAGES C CXX)
 
-if (WIN32)
-    set(CMAKE_INSTALL_LIBDIR $ENV{APPDATA})
-endif()
+find_package(exip CONFIG REQUIRED)
 
-include(${CMAKE_INSTALL_LIBDIR}/lib/cmake/exip/exipTargets.cmake)
+add_library(exi2xml STATIC)
 
-aux_source_directory(. SRC_LIST)
-add_library(exi2xml STATIC ${SRC_LIST})
-target_link_libraries(exi2xml PRIVATE exip)
+target_sources(exi2xml PRIVATE src/XmlWriter.cpp)
 
-target_include_directories(exi2xml
-          PUBLIC 
-          "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>"
-          "$<INSTALL_INTERFACE:include>"
-          )
-
-file(GLOB LIB_INCLUDE_FILES "./*.h")
-
-# install the target and create export-set
-install(TARGETS exi2xml 
-    EXPORT exi2xmlTargets
-    LIBRARY DESTINATION lib
-    ARCHIVE DESTINATION lib
-    RUNTIME DESTINATION bin
-    INCLUDES DESTINATION include
+# define the C++ standard needed to compile this library and make it visible to
+# dependers
+target_compile_features(
+    LabelDemux
+    PUBLIC cxx_std_14
 )
 
-# install header file
-install(FILES ${LIB_INCLUDE_FILES} DESTINATION include)
+target_link_libraries(exi2xml PRIVATE exip)
 
-# generate and install export file
-install(EXPORT exi2xmlTargets
-        FILE exi2xmlTargets.cmake
-        NAMESPACE lcss::
-        DESTINATION ${CMAKE_INSTALL_LIBDIR}/lib/cmake/exi2xml
+target_include_directories(
+    exi2xml
+    PRIVATE src/exi2xml
+    PUBLIC include
+)
+
+include (GNUInstallDirs)
+
+install(TARGETS exi2xml)
+
+install (
+	DIRECTORY include/
+	DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}"
 )
