+++ exi2xml/include/exi2xml/XmlWriter.h
